# docs/swagger.yaml
basePath: /api/v1
definitions:
  auth_service.CreateUserRequest:
    properties:
      age:
        maximum: 130
        minimum: 0
        type: integer
      email:
        maxLength: 100
        minLength: 5
        type: string
      first_name:
        maxLength: 100
        minLength: 3
        type: string
      last_name:
        maxLength: 100
        minLength: 3
        type: string
      password:
        maxLength: 100
        minLength: 10
        type: string
    required:
    - email
    - first_name
    - password
    type: object
  auth_service.LoginDTOResponse:
    properties:
      access_token:
        type: string
      expires_at:
        type: string
      refresh_token:
        type: string
    type: object
  auth_transport_http.CreateUserResponse:
    properties:
      age:
        type: integer
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      updated_at:
        type: string
      version:
        type: integer
    type: object
...
  auth_transport_http.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    type: object
  core_transport_http_response.ErrorResponse:
    properties:
      error:
        example: full error text
        type: string
      message:
        example: short msg
        type: string
    type: object
...
  households_service.RenameHouseholdRequest:
    properties:
      name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - name
    type: object
  households_transport_http.HouseholdDTOResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
      updated_at:
        type: string
      version:
        type: integer
    type: object
...
  households_transport_http.MemberDTOResponse:
    properties:
      email:
        type: string
      first_name:
        type: string
      joined_at:
        type: string
      last_name:
        type: string
      role:
        type: string
      user_id:
        type: string
    type: object
...
  tasks_transport_http.PatchTaskRequest:
    type: object
  tasks_transport_http.TaskDTOResponse:
    properties:
      assigned_to:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      household_id:
        type: string
      id:
        type: string
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
      version:
        type: integer
    type: object
...
host: 127.0.0.1:5050
info:
  contact: {}
  description: Cohesive Application REST-API scheme
  title: Cohesive-Core
  version: "0.1"
paths:
...
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Обновление токена авторизованного пользователя
      parameters:
      - description: RefreshToken тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth_transport_http.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Токен успешно обновлён
          schema:
            $ref: '#/definitions/auth_service.LoginDTOResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "401":
          description: Invalid or expired refresh token
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "429":
          description: Too many requests
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
      summary: Обновление токена
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Создать нового пользователя в системе
      parameters:
      - description: CreateUser тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth_service.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Успешно созданный пользователь
          schema:
            $ref: '#/definitions/auth_transport_http.CreateUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
      summary: Создать пользователя
      tags:
      - auth
...
  /households/{id}:
    delete:
      description: Безвозвратно удалить дом целиком (каскадом удаляются участники,
        инвайты и задачи). Доступно только owner
      parameters:
      - description: ID дома
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Дом удалён
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "403":
          description: Forbidden - only owner can delete household
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "404":
          description: Household not found or not a member
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Удалить дом
      tags:
      - households
    get:
      description: Получить карточку дома. Доступно только участнику дома
      parameters:
      - description: ID дома
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Дом
          schema:
            $ref: '#/definitions/households_transport_http.HouseholdDTOResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "404":
          description: Household not found or not a member
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получить дом
      tags:
      - households
    patch:
      consumes:
      - application/json
      description: Изменить название дома. Доступно owner и admin
      parameters:
      - description: ID дома
        in: path
        name: id
        required: true
        type: string
      - description: RenameHousehold тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/households_service.RenameHouseholdRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Дом переименован
          schema:
            $ref: '#/definitions/households_transport_http.HouseholdDTOResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "403":
          description: Forbidden - role can't rename household
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "404":
          description: Household not found or not a member
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "409":
          description: Version conflict
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Переименовать дом
      tags:
      - households
...
  /households/{id}/members:
    get:
      description: Получить список участников дома с ролями. Доступно любому участнику
      parameters:
      - description: ID дома
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список участников
          schema:
            items:
              $ref: '#/definitions/households_transport_http.MemberDTOResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "404":
          description: Household not found or not a member
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Список участников дома
      tags:
      - households
...
  /households/{id}/tasks/{task_id}:
    delete:
      description: Удалить задачу по ID
      parameters:
      - description: ID дома
        in: path
        name: id
        required: true
        type: string
      - description: ID задачи
        in: path
        name: task_id
        required: true
        type: string
      responses:
        "204":
          description: Задача удалена
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
      summary: Удалить задачу
      tags:
      - tasks
    get:
      description: Получить задачу по ID
      parameters:
      - description: ID дома
        in: path
        name: id
        required: true
        type: string
      - description: ID задачи
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Задача
          schema:
            $ref: '#/definitions/tasks_transport_http.TaskDTOResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
      summary: Получить задачу
      tags:
      - tasks
    patch:
      consumes:
      - application/json
      description: Частично обновить задачу (title, description, status, assigned_to)
      parameters:
      - description: ID дома
        in: path
        name: id
        required: true
        type: string
      - description: ID задачи
        in: path
        name: task_id
        required: true
        type: string
      - description: PatchTask тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/tasks_transport_http.PatchTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Задача обновлена
          schema:
            $ref: '#/definitions/tasks_transport_http.TaskDTOResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "409":
          description: Version conflict
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
      summary: Обновить задачу
      tags:
      - tasks
...
swagger: "2.0"