# internal/features/auth/transport/http/create_user.go
package auth_transport_http

import (
	core_logger "cohesive-core/internal/core/logger"
	core_transport_http_request "cohesive-core/internal/core/transport/http/request"
	core_transport_http_response "cohesive-core/internal/core/transport/http/response"
	auth_service "cohesive-core/internal/features/auth/service"
	"net/http"
)

type CreateUserResponse UserDTOResponse

// CreateUser godoc
// @Summary Создать пользователя
// @Description Создать нового пользователя в системе
// @Tags auth
// @Accept json
// @Produce json
// @Param request body auth_service.CreateUserRequest true "CreateUser тело запроса"
// @Success 201 {object} CreateUserResponse "Успешно созданный пользователь"
// @Failure 400 {object} core_transport_http_response.ErrorResponse "Bad request"
// @Failure 500 {object} core_transport_http_response.ErrorResponse "Internal server error"
// @Router /auth/register [post]
func (h *AuthHTTPHandler) CreateUser(rw http.ResponseWriter, r *http.Request) {
	ctx := r.Context()
	log := core_logger.FromContext(ctx)
	responseHandler := core_transport_http_response.NewHTTPResponseHandler(log, rw)

	log.Debug("invoke CreateUser handler")

	var request auth_service.CreateUserRequest
	if err := core_transport_http_request.DecodeAndValidateRequest(r, &request); err != nil {
		responseHandler.ErrorResponse(err, "failsed to decode and validate HTTP request")

		return
	}

	userDomain, err := h.authService.CreateUser(ctx, request)
	if err != nil {
		responseHandler.ErrorResponse(err, "failed to create user")

		return
	}

	response := CreateUserResponse(userDTOFromDomain(userDomain))
	responseHandler.JSONResponse(response, http.StatusCreated)
}