# internal/features/households/transport/http/rename_household.go
package households_transport_http

import (
	core_errors "cohesive-core/internal/core/errors"
	core_logger "cohesive-core/internal/core/logger"
	core_transport_http_middleware "cohesive-core/internal/core/transport/http/middleware"
	core_transport_http_request "cohesive-core/internal/core/transport/http/request"
	core_transport_http_response "cohesive-core/internal/core/transport/http/response"
	households_service "cohesive-core/internal/features/households/service"
	"fmt"
	"net/http"

	"github.com/google/uuid"
)

// RenameHousehold godoc
// @Summary Переименовать дом
// @Description Изменить название дома. Доступно owner и admin
// @Tags households
// @Accept json
// @Produce json
// @Security ApiKeyAuth
// @Param id path string true "ID дома"
// @Param request body households_service.RenameHouseholdRequest true "RenameHousehold тело запроса"
// @Success 200 {object} households_transport_http.HouseholdDTOResponse "Дом переименован"
// @Failure 400 {object} core_transport_http_response.ErrorResponse "Bad request"
// @Failure 401 {object} core_transport_http_response.ErrorResponse "Unauthorized"
// @Failure 403 {object} core_transport_http_response.ErrorResponse "Forbidden - role can't rename household"
// @Failure 404 {object} core_transport_http_response.ErrorResponse "Household not found or not a member"
// @Failure 409 {object} core_transport_http_response.ErrorResponse "Version conflict"
// @Failure 500 {object} core_transport_http_response.ErrorResponse "Internal server error"
// @Router /households/{id} [patch]
func (h *HouseholdsHTTPHandler) RenameHousehold(rw http.ResponseWriter, r *http.Request) {
	ctx := r.Context()
	log := core_logger.FromContext(ctx)
	responseHandler := core_transport_http_response.NewHTTPResponseHandler(log, rw)

	callerID, ok := core_transport_http_middleware.UserIDFromContext(ctx)
	if !ok {
		responseHandler.ErrorResponse(fmt.Errorf("user id not found in request context"), "internal server errors")

		return
	}

	householdID, err := uuid.Parse(r.PathValue("id"))
	if err != nil {
		responseHandler.ErrorResponse(
			fmt.Errorf("parse household id %q: %v: %w", r.PathValue("id"), err, core_errors.ErrInvalidArgument),
			"invalid household id",
		)

		return
	}

	var request households_service.RenameHouseholdRequest
	if err := core_transport_http_request.DecodeAndValidateRequest(r, &request); err != nil {
		responseHandler.ErrorResponse(err, "failed to decode and validate HTTP request")

		return
	}

	householdWithRole, err := h.householdsService.RenameHousehold(ctx, householdID, callerID, request)
	if err != nil {
		responseHandler.ErrorResponse(err, "failed to rename household")

		return
	}

	response := householdDTOFromDomain(householdWithRole.Household, householdWithRole.Role)
	responseHandler.JSONResponse(response, http.StatusOK)
}